Highlight and Movement restriction
Finished the basic movement base. It is missing the knight movement and the pawn movement in case of an enemy piece being on the sides. I will add the latter when enemy pieces are added into the code.

Got rid of the old highlight functions to the improve version using the BasePiece functions with override.

To fix one of the bugs for now I decided to multiuse the highlight system in the GridManager that was already implemented to set the tiles as not walkable once done. I may change this in the feature to be a function for marking movable tiles that also highlights the posible movable tiles or I may fix it so that it is a separated function.