Basic movement: new classes and prefabs
I decided to further test the options available when using the base class of the pieces to improve readability on the SetInRange() function responsible for the movement and highlight on the tiles.
Now there is an individual script inheriting from the Player Piece class for each piece overriding their movement, and in the case of the horse, adding new functions.
This led to my first ever use of events to change the movement of the pawns after their first move (without including the game manager event that was heavily inspired on a tutorial).
All basic movement of the pieces is completed only missing a few features, e.g. disabling enemy pieces.

This update was done to now test on using generic classes to help some of the functions.