Black pieces, turns and DebugHelper class
Added new prefabs and scriptable objects for the black pieces.

Made some edits to the existing code to accommodate for changing turns in between the white and black pieces as I have decided to make the game 2 players for now until I decide to give a go to an AI. To do this I added checks on the pieces faction to decide which pieces are able to move.

Added the black pieces to the GridGeneration(), few changes made to the SpawnPieces(). So the full board is generated automatically now.

Decided to create a DebugHelper class to store the logs I usually use to debug the classes. I plan to take them out eventually and reuse the class in future projects. I am gonna add generics to it eventually.