New victory state, game reset and pawn/king
Added the victory game state to the game once one of the kings is disabled (eaten/taken out).

During the victory state is possible to rematch, in order to do this a few changes were made in the game manager to check if the grid had already been generated when the board generation game state is called. For now all pieces are stored in a list when they are generated and deleted before resetting the board by generating them again opposed to reusing them for a quicker solution, I may change that in the future to object pooling.

Menu manager was changed to possibly add more to the victory state.

Pawns were updated to fix their movement to match an actual pawn movement.

Known bugs:
There is a null warning when the kings are destroyed/disabled once the game is stopped as they call for the menu manager. Doesn't affect the actual game. I will fix it by the next backup

Also fixed the mistake of using enable when eating pieces instead of disabling the game object.