TileBehaviour and Castling/UnitManager
Separated the Tile behaviour that handled all the piece movement to a separated component to keep a more specific class. Now TileBehaviour deals with all interactions done through the tile while the Tile only contains the relevant information like appearance and the occupying unit.

Finally added the castling mechanic to the game (adding the Castling component to the UnitManager). After looking more deeply into it I realised that castling was done through the king opposed to the tower doing it. In my mind castling was done through the rook/tower, it made sense to me... Apparently not. I was already done by the time I realised, so I decided to make it possible both ways. Thanks to this I learned better about events and how to use them. Left the events on the tower and king in case I decide to use it for something eventually.

Added an extra tile of movement for castling to the tower, but now it looks weird... may take it out later.