PawnUpgrade and future object pooling
I worked a bit on a base for the pawn upgrade. So far I have the transition from the turn to the PawnUpgrade state done and working. I tried to check if changing the scriptable object would for some reason work. It didn't.

I was thinking about maybe going into the scriptable script and make it do the change on one of the methods like OnEnable(), but I think I will simply add object pooling from my other project to use it both for upgrading the pawn and resetting the board. In the future I will look more into scriptable objects for this type of use.

Next time Object Pooling should be added and PawnUpgrade prototype finished.