Fixed Castling, cam rotation
Changed castling to store the castling tiles on generation. Added new checks to avoid the event getting called after any of the pieces are no longer capable of castling.

Added a new cam rotation script to handle the turn changed making the camera focus on the current factions side. Will probably move the EvaluateSlerpPoints to the general ObjectLerper class as a virtual and change it in cam rotation.